fix bug for run scripts

+++ scripts/linux/publish_docker.simple_kv.sh
@@ -9,14 +9,15 @@ applist="meta replica client client.perf.test"
 echo $applist > $t_dir/applist
 
 cp $b_dir/bin/dsn.replication.simple_kv/dsn.replication.simple_kv $t_dir/simple_kv
-cp $b_dir/bin/dsn.replication.simple_kv/config.ini $t_dir/
 cp $scripts_dir/deploy/configtool $t_dir/
 cp $DSN_ROOT/lib/libdsn.core.so $t_dir/ 
 
-echo "please customize your config.ini"
-sleep 5
-vim $t_dir/config.ini
-
+if [ ! -f $t_dir/config.ini ];then
+    cp $b_dir/bin/dsn.replication.simple_kv/config.ini $t_dir/
+    echo "please customize your config.ini"
+    sleep 5
+    vim $t_dir/config.ini
+fi
 sed -e "s/{{ placeholder\['deploy_name'\] }}/simple_kv/g" $scripts_dir/deploy/run.sh > $t_dir/run.sh
 
 chmod a+x $t_dir/run.sh
@@ -27,9 +28,8 @@ COPY simple_kv /home/rdsn/
 COPY config.ini /home/rdsn/
 COPY run.sh /home/rdsn/
 RUN mkdir -p /home/rdsn/data
-
+RUN chown -R rdsn:rdsn /home/rdsn 
 ENV LD_LIBRARY_PATH=/home/rdsn/lib
-
 CMD ["./run.sh"]
 EOF
